+++ staramr/blast/pointfinder/PointfinderDatabaseInfo.py
@@ -37,11 +37,12 @@ class PointfinderDatabaseInfo:
 
         with open(file) as f:
             line = f.readline()
-        
+
         line = line.lstrip("#")
         column_names = line.split()
 
-        pointfinder_info = pd.read_csv(file, sep='\t', index_col=False, comment='#', header=None, names=column_names)
+        pointfinder_info = pd.read_csv(file, sep=r'\t|\s{4,}', index_col=False, comment='#', header=None, names=column_names, engine='python')
+        pointfinder_info["PMID"] = pointfinder_info["PMID"].astype(str)
 
         return cls(pointfinder_info, file)
 
@@ -53,7 +54,7 @@ class PointfinderDatabaseInfo:
         :return: A new PointfinderDatabaseInfo.
         """
         return cls(database_info_dataframe)
-    
+
     @staticmethod
     def to_codons(regex_match):
         # Sometimes, the regex will match a string with a comma and return multiple matches.
@@ -107,6 +108,14 @@ class PointfinderDatabaseInfo:
         & (table['Gene_ID'].str.contains('16S') == False)
         & (table['Gene_ID'].str.contains('23S') == False), "Res_codon"].str.replace('[A-Z,]+', self.to_codons, regex=True)
 
+        # We need to correct for the mismatch between promter's FASTA record ID
+        # and the associated gene ID in the resistens-overview.txt file. For example:
+        # FASTA Record ID: >ampC-promoter_1_CP037449.1
+        # Gene ID: ampC_promoter_size_53bp
+        # The gene ID seems to be derived from the file name (ampC-promoter-size-53bp.fsa),
+        # although that still contains mismatching hyphens and underscores.
+        table[['Gene_ID']] = table[['Gene_ID']].replace({r'(.+promoter).+' : r'\1'}, regex=True)
+
     def _get_resistance_codon_match(self, gene, codon_mutation):
         table = self._pointfinder_info
 
@@ -120,7 +129,7 @@ class PointfinderDatabaseInfo:
                         # so we need to convert to nucleotide coordinates before making the comparison.
                     & (table['Ref_codon'] == codon_mutation.get_database_amr_gene_mutation())
                     & (table['Res_codon'].str.contains(codon_mutation.get_input_genome_mutation(), regex=False))]
-        
+
         # We need to handle codon insertions as a special case:
         # Pointfinder mis-reports the position of codon insertions. For example:
         # ref:     ACG --- ACG